Fix get-set coments, get-set likes
